+++ src/components/views/settings/encryption/RecoveryPanel.tsx
@@ -13,8 +13,6 @@ import { SettingsSection } from "../shared/SettingsSection";
 import { _t } from "../../../../languageHandler";
 import { useMatrixClientContext } from "../../../../contexts/MatrixClientContext";
 import { SettingsHeader } from "../SettingsHeader";
-import { accessSecretStorage } from "../../../../SecurityManager";
-import { SettingsSubheader } from "../SettingsSubheader";
 import { useAsyncMemo } from "../../../../hooks/useAsyncMemo";
 
 /**
@@ -85,37 +83,6 @@ export function RecoveryPanel({ onChangeRecoveryKeyClick }: RecoveryPanelProps):
             data-testid="recoveryPanel"
         >
             {content}
-        </SettingsSection>
-    );
-}
-
-interface RecoveryPanelOutOfSyncProps {
-    /**
-     * Callback for when the user has finished entering their recovery key.
-     */
-    onFinish: () => Promise<void>;
-}
-
-/**
- * This component is shown when the user secrets are not cached and the recovery key needs to be enter.
- */
-export function RecoveryPanelOutOfSync({ onFinish }: RecoveryPanelOutOfSyncProps): JSX.Element {
-    return (
-        <SettingsSection
-            legacy={false}
-            heading={_t("settings|encryption|recovery|title")}
-            subHeading={
-                <SettingsSubheader
-                    label={_t("settings|encryption|recovery|description")}
-                    state="error"
-                    stateMessage={_t("settings|encryption|recovery|key_storage_warning")}
-                />
-            }
-            data-testid="recoveryPanel"
-        >
-            <Button size="sm" kind="primary" Icon={KeyIcon} onClick={async () => await accessSecretStorage(onFinish)}>
-                {_t("settings|encryption|recovery|enter_recovery_key")}
-            </Button>
         </SettingsSection>
     );
 }
